% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{add_chilled_water_eui}
\alias{add_chilled_water_eui}
\title{Add chilled water eui}
\usage{
add_chilled_water_eui()
}
\description{
This function adds chilled water eui
}
\examples{
add_chilled_water_eui()
}
\keyword{db}
\keyword{modify}
