% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{add_quality_tag_energy}
\alias{add_quality_tag_energy}
\title{Add energy sanity check}
\usage{
add_quality_tag_energy()
}
\description{
This function produces EUAS_yearly table, with all columns aggregated to
yearly
}
\examples{
add_quality_tag_energy()
}
\keyword{quality}
\keyword{tag}
