% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{backup_table}
\alias{backup_table}
\title{Back up a table to a csv file}
\usage{
backup_table(dbname, tablename)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{tablename}{required, the name string of the table to view}
}
\description{
This function backs up the specified table in db_build_temp_csv, with a timestamp attached to the filename
}
\examples{
backup_table(dbname="all", tablename="EUAS_address")
}
\keyword{backup}
\keyword{table}
