% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check duplicate of table}
\usage{
check_duplicates(dbname, tablename, groupby_vars)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{tablename}{required, the name string of the table to view}

\item{groupby_vars}{required, columns to group by at}
}
\description{
This function checks duplicates of a table, group by columns
}
\examples{
backup_table(dbname="all", tablename="EUAS_address")
}
\keyword{backup}
\keyword{table}
