% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{connect}
\alias{connect}
\title{Get the connection object of sqlite database with name "dbname"}
\usage{
connect(dbname, path)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{path}{optional, the path to .db file, default is "csv_FY/db/"}
}
\description{
This function returns a connector to dbname
}
\examples{
connect("all")
}
\keyword{connection}
\keyword{sqlite}
