% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{correct_city_name}
\alias{correct_city_name}
\title{Correct city name}
\usage{
correct_city_name()
}
\description{
This function overwrites the city name error, or unify their names
}
\examples{
correct_city_name()
}
\keyword{city}
\keyword{name}
