% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{drop_table_from_db}
\alias{drop_table_from_db}
\title{Drop a table from .db}
\usage{
drop_table_from_db(dbname, tablename)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{tablename}{required, the name string of the table to view}
}
\description{
This function deletes a table from database
}
\examples{
drop_table_from_db("all", "EUAS_type_")
}
\keyword{drop}
\keyword{table}
