% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{get_all_tables}
\alias{get_all_tables}
\title{Get all tables in a sqlite .db file}
\usage{
get_all_tables(dbname, path)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{path}{optional, the path to .db file, default is "csv_FY/db/"}
}
\description{
This function returns a vector of all tables in the database
}
\examples{
get_all_tables("all")
}
\keyword{all_tables}
\keyword{sqlite}
