% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{get_buildings}
\alias{get_buildings}
\title{Get list of buildings}
\usage{
get_buildings(region, buildingType, year, category)
}
\arguments{
\item{region}{optional, region number}

\item{buildingType}{optional, building type}

\item{year}{optional, restrict to data with fiscal year = year}

\item{category}{optional, restrict to data with category in a vector of
categories, c(a vector of categories, e.g. "A", "I")}
}
\description{
This function returns a vector of building id's
}
\examples{
get_buildings()
}
\keyword{all}
\keyword{buildings}
\keyword{get}
