% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{get_count}
\alias{get_count}
\title{Get the count of buildings with region, category (A, C, I), type (office, etc.), and year filters applied}
\usage{
get_count(region, category, type, fOrC)
}
\arguments{
\item{region}{optional, a string vector of region numbers, or a single
region number}

\item{category}{optional, a subset of A, B, C, D, E, I}

\item{type}{optional, building type}

\item{fOrC}{optional, specify one of "F" (fiscal year) or "C" (calendar year), default to "F"}

\item{year}{optional, a double vector of years, or a single year}
}
\description{
This function returns the count group by region, category, and type, if any
of the parameters are specified, they will be used in filtering the results
}
\examples{
get_count(1, c("A", "C", "I"), "office", "F")
}
\keyword{count}
\keyword{query}
