% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{get_euas_buildings}
\alias{get_euas_buildings}
\title{Get all buildings}
\usage{
get_euas_buildings()
}
\description{
This function returns a data frame with one column, Building_Number,
containing all buildings in the EUAS database
}
\examples{
get_euas_buildings()
}
\keyword{all}
\keyword{buildings}
\keyword{get}
