% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{get_eui_by_year}
\alias{get_eui_by_year}
\title{Get eui by fiscal year}
\usage{
get_eui_by_year(fOrC)
}
\arguments{
\item{fOrC}{optional, "F" for fiscal year, "C" for calendar year}
}
\description{
This function produces EUAS_yearly table, with all columns aggregated to
yearly
}
\examples{
get_eui_by_year("F")
}
\keyword{by}
\keyword{eui}
\keyword{year}
