% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{get_heating_cooling_eui}
\alias{get_heating_cooling_eui}
\title{compute heating and cooling energy consumption per square foot}
\usage{
get_heating_cooling_eui(debugFlag = FALSE)
}
\description{
This function computes the heating and cooling energy use per square foot for
each building. Assume heating fuels include gas, oil, and steam; cooling fuel
include electricity and chilled water. If there's zero gas + oil + steam for
a year, assume electricity heating.
}
\examples{
get_heating_cooling_eui
}
\keyword{cooling}
\keyword{eui}
\keyword{heating}
