% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{get_lat_lon_df}
\alias{get_lat_lon_df}
\title{Get latitude longitude}
\usage{
get_lat_lon_df(path, building)
}
\arguments{
\item{path}{to the all.db file, default is "csv_FY/db/"}

\item{building}{optional, if supplied, only this building's data is returned}
}
\description{
This function get latitude longitude of EUAS buildings
}
\examples{
get_lat_lon_df(building="XXXXXXXX")
}
\keyword{latlon}
