% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{get_unique_value_column}
\alias{get_unique_value_column}
\title{Get unique values from a column}
\usage{
get_unique_value_column(dbname, tablename, path, col)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{tablename}{required, the name string of the table to view}

\item{path}{optional, the path to .db file, default is "csv_FY/db/"}

\item{col}{required, the column to get unique value of}
}
\description{
This function returns a table from database
}
\examples{
get_unique_value_column(dbname="all", tablename="EUAS_type_recode", col="Building_Type")
}
\keyword{head}
\keyword{sqlite}
\keyword{table}
\keyword{view}
