% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{join_type_and_energy}
\alias{join_type_and_energy}
\title{Join EUAS_monthly and EUAS_type}
\usage{
join_type_and_energy()
}
\description{
This function joins EUAS_monthly and EUAS_type
}
\examples{
join_type_and_energy()
}
\keyword{drop}
\keyword{table}
