% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{main_db_build}
\alias{main_db_build}
\title{Join EUAS_monthly and EUAS_type}
\usage{
main_db_build()
}
\description{
This function joins EUAS_monthly and EUAS_type
}
\examples{
main_db_build()
}
\keyword{drop}
\keyword{table}
