% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{read_table_from_db}
\alias{read_table_from_db}
\title{Read table from db}
\usage{
read_table_from_db(dbname, tablename, path, cols, building)
}
\arguments{
\item{dbname}{required, the name string of the database, e.g. "all.db" has name "all"}

\item{tablename}{required, the name string of the table to view}

\item{path}{optional, the path to .db file, default is "csv_FY/db/"}

\item{cols}{optional, the columns to select}

\item{building}{optional, only read data for this building}
}
\description{
This function returns a table from database
}
\examples{
view_head_of_table(dbname="all", tablename="EUAS_type")
}
\keyword{head}
\keyword{sqlite}
\keyword{table}
\keyword{view}
