% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{recode_euas_type}
\alias{recode_euas_type}
\title{Recode building type}
\usage{
recode_euas_type()
}
\description{
This function recode different variations of building types, e.g. Office and
Office Building is combined into the Office category
}
\examples{
recode_euas_type()
}
\keyword{building}
\keyword{recode}
\keyword{type}
