% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{recode_state_abbr}
\alias{recode_state_abbr}
\title{Recode state to abbreviation}
\usage{
recode_state_abbr()
}
\description{
This function converts state full name in EUAS data to state abbreviation, as
the current state in the data is a mixture of abbreviation and full name
}
\examples{
recode_state_abbr()
}
\keyword{abbr}
\keyword{state}
\keyword{to}
