% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{remove_old_energy_data}
\alias{remove_old_energy_data}
\title{Remove old energy data based on index}
\usage{
remove_old_energy_data()
}
\description{
This function removes the old energy data
}
\examples{
remove_old_energy_data()
}
\keyword{drop}
\keyword{old}
