% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{unify_euas_type}
\alias{unify_euas_type}
\title{Unify building type}
\usage{
unify_euas_type()
}
\description{
This function compiles building type information to a data frame containing
Building_Number, Building_Type, data_source. The input tables are
from a database other_input. There are currently three sources of
building type: their order of importance are
"euas_database_of_buildings_cmu::[GSA Property Type]" >
"PortfolioManager_sheet0_input::Self-Selected_Primary_Function" >
"Entire_GSA_Building_Portfolio_input::Predominant_Use")))
}
\examples{
unify_euas_type()
}
\keyword{building}
\keyword{type}
\keyword{unify}
