% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{compute_eui}
\alias{compute_eui}
\title{Add chilled water eui}
\usage{
compute_eui(df, energy_input, eui_output, sqftcol, mult)
}
\description{
This function adds chilled water eui
}
\examples{
compute_eui()
}
\keyword{db}
\keyword{modify}
