% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{unify_euas_type}
\alias{unify_euas_type}
\title{Unify building type}
\usage{
unify_euas_type()
}
\description{
This function compiles building type information to a data frame with column.
Assuming the required tables are dumped into a database other_input
}
\examples{
unify_euas_type()
}
\keyword{building}
\keyword{type}
\keyword{unify}
